pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "wpipe"
version = "1.5.3"
authors = [
    { name = "William Steve Rodriguez Villamizar", email = "wisrovi.rodriguez@gmail.com" }
]
description = "Pipeline library with API integration for task orchestration and execution tracking"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.9"
keywords = [
    "pipeline",
    "workflow",
    "tasks",
    "orchestration",
    "api",
    "worker",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Intended Audience :: Developers",
    "Development Status :: 5 - Production/Stable",
]
dependencies = [
    "requests>=2.31.0",
    "loguru>=0.7.0",
    "pandas>=2.0.0",
    "pyyaml>=6.0.1",
    "tqdm>=4.66.0",
    "rich>=13.7.0",
    "pydantic>=2.0.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "wsqlite>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]
docs = [
    "sphinx>=7.2.0",
    "sphinx-rtd-theme>=2.0.0",
    "sphinx-copybutton>=0.5.0",
    "myst-parser>=2.0.0",
    "sphinx-sitemap>=2.5.0",
    "sphinx-design>=0.5.0",
]

[project.urls]
Homepage = "https://github.com/wisrovi/wpipe"
Documentation = "https://wpipe.readthedocs.io/en/latest/"
Repository = "https://github.com/wisrovi/wpipe"
Issues = "https://github.com/wisrovi/wpipe/issues"
Changelog = "https://github.com/wisrovi/wpipe/blob/main/CHANGELOG.md"

[tool.hatch.build.targets.sdist]
include = [
    "/wpipe",
    "/examples/01_basic_pipeline",
    "/examples/02_api_pipeline",
    "/examples/11_checkpointing",
    "/examples/12_timeouts",
    "/examples/13_type_hinting",
    "/examples/14_resource_monitor",
    "/examples/15_export",
    "/examples/16_parallelism",
    "/test",
]
exclude = [
    "**/*.db",
    "**/*.log",
    "**/*.zip",
    "**/__pycache__",
    "**/configs",
    "**/logs",
    "**/*.yaml",
]

[tool.hatch.build.targets.wheel]
packages = ["wpipe"]
exclude = [
    "**/*.db",
    "**/*.log",
    "**/*.zip",
    "**/__pycache__",
]

[tool.pytest.ini_options]
testpaths = ["test"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["wpipe"]
branch = true
omit = [
    "*/tests/*",
    "*/examples/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
show_missing = true
fail_under = 72

[tool.black]
line-length = 88
target-version = ["py39", "py310", "py311"]
include = '\.pyi?$'

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long
    "B008",  # do not perform function calls in argument defaults
]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
